Ya imprime solo los campos solicitados de un registro. Por ejemplo se puede hacer: select persona.nombre from persona;
Tambien se comprobo que en todos los switchs por tipos de datos no falte ningun tipo.

+++ CapaIndices/Hash/Hash.cpp
@@ -339,7 +339,7 @@
 			*/
 			
 			// Algoritmo djb2
-			unsigned long hash = 5381;
+			int hash = 5381;
 	        int c;
 	        
 	        c = *clave++;
